Dockerfile, UI Changes
Updated the Dockerfile in hopes to add the ability to add the docker to Unraid, however I am not having much success. Right now, Dockers do work, but not well and there is no configuration for it.

Updated the UI a bit more to better support mobile and give a cleaner look.